+++ framework/domain/src/main/java/is/codion/framework/domain/entity/ImmutableEntity.java
@@ -30,7 +30,7 @@ import java.io.Serializable;
 import java.util.HashMap;
 import java.util.Map;
 
-import static is.codion.framework.domain.entity.DefaultKey.serializerForDomain;
+import static is.codion.framework.domain.entity.EntitySerializer.serializerForDomain;
 
 final class ImmutableEntity extends DefaultEntity implements Serializable {